Bug 499 (TextField copy/paste bug) patch
See http://echo.nextapp.com/site/node/6578
This is a not-terribly-well tested patch for Issue 499, the textfield cut/paste bug.
It adds a fairly significant (albeit very small) change to Client.js and the client-side
component peer architecture, i.e., the firing of renderBlur() events to peers. So far this
has been the only way I've found to prevent the issue, as browsers handle the textfield
DOM onchange in a fairly annoying way.